Created new client/auth
