+++ edgegrid/client/client.go
@@ -1,0 +1,7 @@
+package client
+
+// A Client implements the base client request and response handling
+// used by all service clients.
+type Client struct {
+	request string
+}
