+++ service/netlistv2/service.go
@@ -1,1 +1,12 @@
 package netlistv2
+
+import (
+	"github.com/apiheat/go-edgegrid/edgegrid/client"
+)
+
+// Netlistv2 provides the API operation methods for making requests to
+// Akamai Netlistv2. See this package's package overview docs
+// for details on the service.
+type Netlistv2 struct {
+	*client.Client
+}
